+++ packages/cli/__tests__/zosjobs/__system__/submit/stdin/__scripts__/submit_valid_stdin.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/env bash
+#!/bin/bash
 # pass the JCL as stdin to this script
 cat "$1" | zowe zos-jobs submit stdin
 exit $?