+++ packages/cli/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to the Zowe CLI package will be documented in this file.
 
 ## Recent Changes
-- BugFix: Updated example on zowe profiles create zosmf-profile command. [#1152](https://github.com/zowe/zowe-cli/issues/1152)
+- BugFix: Updated example for `zowe profiles create zosmf-profile` command. [#1152](https://github.com/zowe/zowe-cli/issues/1152)
 
 ## `7.6.1`