+++ packages/cli/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to the Zowe CLI package will be documented in this file.
 
+## Recent Changes
+- Enhancement: Added `ZosFilesCreateOptions.alcunit` option to pds definition. [#1203](https://github.com/zowe/zowe-cli/issues/1203)
+
 ## `7.6.1`
 
 - BugFix: Updated `ssh2` dependency to fix "Received unexpected packet type" error on SSH commands. [#1516](https://github.com/zowe/zowe-cli/issues/1516)