+++ packages/cli/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the Zowe CLI package will be documented in this file.
 
-## Recent Changes
+## `7.6.2`
 
 - BugFix: Updated `minimatch` and `keytar` dependencies for technical currency.
 - BugFix: Updated example for `zowe profiles create zosmf-profile` command. [#1152](https://github.com/zowe/zowe-cli/issues/1152)